+++ CHANGELOG.md
@@ -7,6 +7,10 @@
   useful summary for people upgrading their application, not a replication
   of the commit log.
 
+## Unreleased
+
+* Add priority breadcrumb ([PR #1501] (https://github.com/alphagov/govuk_publishing_components/pull/1501))
+
 ## 21.48.0
 
 * Expand js classes code ([PR #1513](https://github.com/alphagov/govuk_publishing_components/pull/1513))